requirements.txt
astroid==2.5.3
greenlet==1.0.0
h5py==3.2.1
imutils==0.5.4
isort==5.8.0
jedi==0.18.0
joblib==1.0.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
msgpack==1.0.2
neovim==0.3.1
numpy==1.20.2
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
parso==0.8.2
progressbar2==3.53.1
pylint==2.7.4
pynvim==0.4.3
python-utils==2.5.6
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.2
wrapt==1.12.1
